+++ versions.tf
@@ -2,8 +2,8 @@ terraform {
   required_providers {
     cloudflare = {
       source  = "cloudflare/cloudflare"
-      version = "~> 2.0"
+      version = "~> 3.0"
     }
   }
-  required_version = ">= 0.15"
+  required_version = ">= 1.1"
 }